tithek thevideo add not found msg

+++ mediathek/localhoster/hoster.sh
@@ -193,7 +193,6 @@ thevideo()
 	$BIN $CMD/thevideo.py $INPUT > $TMP/cache.$FILENAME.1
 
 	videocode=`cat $TMP/cache.$FILENAME.1`
-
 	rm -f $TMP/cache.hoster.$hoster.* > /dev/null 2>&1
 
 	STREAMLIST="$TMP/$TYPE.$hoster.$FILENAME.streamlist"
@@ -204,19 +203,25 @@ thevideo()
 	$curlbin --data '{}' https://vev.io/api/serve/video/$videocode -o $TMP/cache.$FILENAME.2
 	cat $TMP/cache.$FILENAME.2 | sed 's/{/\n/g' | sed 's/,/\n/g' | grep "/stream/" >$TMP/cache.$FILENAME.3
 
-	while read -u 3 ROUND; do
-		TITEL=`echo $ROUND | cut -d'"' -f2`
-		URL=`echo $ROUND | cut -d'"' -f4`
+	#{"code":400,"message":"invalid video specified","errors":[]}
+	if [ `cat $TMP/cache.$FILENAME.2 | grep '"code":400,' | wc -l` -eq 1 ];then
+		ERROR=`cat $TMP/cache.$FILENAME.2 | sed -nr 's/.*"message":"([^"]+)".*/\1/p'`
+		echo "errormsg=$ERROR"
+	else
+		while read -u 3 ROUND; do
+			TITEL=`echo $ROUND | cut -d'"' -f2`
+			URL=`echo $ROUND | cut -d'"' -f4`
 
-		if [ ! -z "$URL" ];then
-			echo "$URL" | tr -d '\n' | tr -d '\r' >> $STREAMLIST
-			echo "" >> $STREAMLIST
-		fi
+			if [ ! -z "$URL" ];then
+				echo "$URL" | tr -d '\n' | tr -d '\r' >> $STREAMLIST
+				echo "" >> $STREAMLIST
+			fi
 
 
-	done 3<$TMP/cache.$FILENAME.3
-	rm $TMP/cache.$FILENAME.* > /dev/null 2>&1
-	echo $STREAMLIST
+		done 3<$TMP/cache.$FILENAME.3
+		rm $TMP/cache.$FILENAME.* > /dev/null 2>&1
+		echo $STREAMLIST
+	fi
 }
 
 movshare()
